refactor(video-player): 再生システムの純粋ロジックを lib/ に分離
コンポーネント worker にシステムロジックが混ざっていたのを整理。
- src/lib/playback.ts: computeCurrentTime / isClockOverrun / formatTime（純関数）
- src/lib/youtube.ts: extractVideoId / thumbnailUrl（youtube.ts から移動）
- controls/playlist は lib を import するだけに。重複していた fmt と
  共有時計計算を一本化（挙動は不変）。

screen は元々ステートレスな「表示・実行」専任で、isPlaying/進捗などの
状態は controls が一元保持（今回の整理で境界がより明確に）。
